added media table to database, bug fix

+++ app/src/main/java/org/nuclearfog/twidda/backend/api/mastodon/impl/MastodonTrend.java
@@ -43,8 +43,8 @@ public class MastodonTrend implements Trend {
 
 
 	@Override
-	public int getLocationId() {
-		return -1;
+	public long getLocationId() {
+		return -1L;
 	}
